+++ src/xrl/agents/policies/genetic_rl.py
@@ -38,25 +38,7 @@ model_name = lambda training_name : PATH_TO_OUTPUTS + training_name + "_model.pt
 
 # TODO: Fix serialization problem and decomment after,
 # look at other genetic_rl.py file for details!
-from xrl.genetic_rl import policy_net
-# define policy network for genetic algo
-#class policy_net(nn.Module):
-#    def __init__(self, input, hidden, actions, make_hidden = True): 
-#        super(policy_net, self).__init__()
-#        # should make one hidden layer
-#        self.make_hidden = make_hidden
-#
-#        if self.make_hidden:
-#            self.h = nn.Linear(input, hidden)
-#            self.out = nn.Linear(hidden, actions)
-#        else:
-#            self.out = nn.Linear(input, actions)
-#
-#
-#    def forward(self, x):
-#        if self.make_hidden:
-#            x = F.relu(self.h(x))
-#        return F.softmax(self.out(x), dim=1)
+from xrl.agents.policies.policy_model import policy_net
 
 
 def init_weights(m):
